requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pyserial>=3.5
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.26
pytest>=7.4
